+++ rest-server/auth-server/src/main/scala/com/actionml/authserver/dal/mongo/UsersDaoImpl.scala
@@ -1,4 +1,20 @@
 package com.actionml.authserver.dal.mongo
+/*
+ * Copyright ActionML, LLC under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * ActionML licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 import com.actionml.authserver.dal.UsersDao
 import com.actionml.authserver.exceptions.UserNotFoundException
@@ -7,7 +23,6 @@ import com.typesafe.scalalogging.LazyLogging
 import org.mongodb.scala._
 import org.mongodb.scala.model.Filters._
 import scaldi.{Injectable, Injector}
-
 import scala.concurrent.{ExecutionContext, Future}
 
 class UsersDaoImpl(implicit inj: Injector) extends UsersDao with MongoSupport with Injectable with LazyLogging {
