+++ rest-server/conf/application.conf
@@ -15,4 +15,6 @@ mongo {
   host = ${?MONGO_HOST}
   port = 27017
   port = ${?MONGO_PORT}
+  sharedDB = "harness-shared"
+  sharedDB = ${?SHARED_DB_NAME}
 }
