+++ rest-server/server/src/main/scala/com/actionml/router/http/routes/EventsRouter.scala
@@ -10,7 +10,7 @@ import com.actionml.authserver.Roles.event
 import com.actionml.authserver.directives.AuthorizationDirectives
 import com.actionml.authserver.service.AuthorizationService
 import com.actionml.authserver.services.AuthServerProxyService
-import com.actionml.router.config.AppConfig
+import com.actionml.core.config.AppConfig
 import com.actionml.router.service._
 import io.circe.Json
 import scaldi.Injector
